requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pyyaml>=6.0
cachetools>=5.3.0
filelock>=3.13.0
soundfile>=0.12.0
python-dotenv>=1.0.0
pytest>=8.0.0
